+++ app/src/main/java/com/achievers/data/source/categories/CategoriesCursor.java
@@ -13,11 +13,11 @@ import static com.achievers.provider.AchieversContract.Categories.INDEXES.CATEGO
 public class CategoriesCursor {
 
     public static Category from(Cursor cursor) {
-        int id = cursor.getInt(CATEGORY_ID);
+        Long id = cursor.getLong(CATEGORY_ID);
         String title = cursor.getString(CATEGORY_TITLE);
         String description = cursor.getString(CATEGORY_DESCRIPTION);
         String imageUrl = cursor.getString(CATEGORY_IMAGE_URL);
-        Integer parentId = cursor.getInt(CATEGORY_PARENT_ID);
+        Long parentId = cursor.getLong(CATEGORY_PARENT_ID);
         if (parentId == 0) parentId = null;
 
         return new Category(id, title, description, imageUrl, parentId);
