+++ app/src/main/java/com/achievers/generator/CategoriesGenerator.java
@@ -13,7 +13,7 @@ public class CategoriesGenerator implements Generator<Category> {
     private static final int InitialRow = 0;
     private static final int InitialCol = 0;
     private static final int InitialId = 0;
-    private static final Integer RootParent = null;
+    private static final Long RootParent = null;
     private static final String RandomImagePath = "https://unsplash.it/500/500/?random&a=%d";
 
     private final Faker mFaker;
@@ -23,7 +23,7 @@ public class CategoriesGenerator implements Generator<Category> {
     }
 
     @Override
-    public Category single(int id, Integer parentId) {
+    public Category single(long id, Long parentId) {
         return generateCategory(id, parentId);
     }
 
@@ -41,10 +41,10 @@ public class CategoriesGenerator implements Generator<Category> {
             int[] sizes,
             int row,
             int col,
-            Integer parentId,
+            Long parentId,
             List<Category> generated) {
 
-        int id = getId(generated);
+        long id = getId(generated);
         generated.add(generateCategory(id, parentId));
 
         if (sizes.length > row + 1) {
@@ -58,14 +58,14 @@ public class CategoriesGenerator implements Generator<Category> {
         return generated;
     }
 
-    private int getId(List<Category> categories) {
+    private long getId(List<Category> categories) {
         int size = categories.size();
-        int lastId = size > 0 ? categories.get(size - 1).getId() : 0;
+        long lastId = size > 0 ? categories.get(size - 1).getId() : 0;
 
         return lastId + 1;
     }
 
-    private Category generateCategory(int id, Integer parentId) {
+    private Category generateCategory(long id, Long parentId) {
         String title = mFaker.lorem.word();
         String description = mFaker.lorem.sentence(5);
         String imageUrl =  String.format(
